Change Timer component props
Change Timer component props so it uses timestamp in units instead of days,
hours, minutes and seconds.
